refactor: Move queue-related code to a new queue.ts file for clarity
