+++ src/index.ts
@@ -1,57 +1,11 @@
 import 'dotenv/config'
 import express from 'express'
-import { createBullBoard } from '@bull-board/api'
-import { BullMQAdapter } from '@bull-board/api/bullMQAdapter'
-import { ExpressAdapter } from '@bull-board/express'
-import { Queue } from 'bullmq'
+import { serverAdapter, queues, cancelActiveJobs } from './queue'
 
 import discord from './discord'
 import apiRouter from './api'
 
 // start ui
-const serverAdapter = new ExpressAdapter()
-serverAdapter.setBasePath('/admin/queues')
-
-async function cancelActiveJobs(queue: Queue) {
-  const jobCounts = await queue.getActiveCount()
-  if (jobCounts > 0) {
-    const activeJobs = await queue.getActive()
-    for (const job of activeJobs) {
-      try {
-        await job.moveToFailed(new Error('Cancelling active job'), queue.token)
-        // await job.retry()
-      } catch (e) {
-        console.error(e)
-      }
-    }
-  }
-}
-
-const workers = [
-  'downloadPDF',
-  'pdf2Markdown',
-  'splitText',
-  'indexParagraphs',
-  'searchVectors',
-  'precheck',
-  'guessWikidata',
-  'extractEmissions',
-  'followUp',
-  'checkDB',
-  'saveToAPI',
-]
-
-const queues = initializeQueues(workers, process.env.REDIS_URL)
-
-createBullBoard({
-  queues: queues.map((queue) => new BullMQAdapter(queue)),
-  serverAdapter: serverAdapter,
-  options: {
-    uiConfig: {
-      boardTitle: 'Klimatkollen',
-    },
-  },
-})
 
 const app = express()
 discord.login()
